fix: handle float hs_call_duration + unblock snapshots after backfill
- Parse hs_call_duration with float() instead of int() to handle
  HubSpot values like "8224539.000000001"
- Reset front_deal_triggered_at after backfill_deal_context completes
- Reduce default workers from 10 to 3 to avoid Azure rate limits

+++ src/pipelines/sync_deal_context/run.py
@@ -512,7 +512,7 @@ def run(deal_uuid: str, hs_deal_id: str, *, owners: dict[str, dict] | None = Non
             sub = get_subteam(owner_email) if owner_email else None
 
             duration_ms = p.get("hs_call_duration")
-            duration_s = int(int(duration_ms) / 1000) if duration_ms else None
+            duration_s = int(float(duration_ms) / 1000) if duration_ms else None
             duration_min = round(duration_s / 60) if duration_s else 0
 
             fecha = _parse_date(p.get("hs_timestamp"))
@@ -589,7 +589,7 @@ def run(deal_uuid: str, hs_deal_id: str, *, owners: dict[str, dict] | None = Non
             owner_info = owners.get(owner_id, {})
             owner_name = owner_info.get("name", owner_info.get("email", "?"))
             duration_ms = p.get("hs_call_duration")
-            duration_min = round(int(int(duration_ms) / 1000) / 60) if duration_ms else 0
+            duration_min = round(int(float(duration_ms) / 1000) / 60) if duration_ms else 0
             date = p.get("hs_timestamp") or ""
             items.append((date, "context", _format_call_context(
                 hs_id, p, owner_name or "?", duration_min
